feat: centralize character movement broadcasting and state management within the player client
