+++ game/src/packets/from_client/move_to_location.rs
@@ -1,13 +1,8 @@
 use anyhow::bail;
-use crate::packets::to_client::CharMoveToLocation;
 use bytes::BytesMut;
-use kameo::actor::ActorRef;
 use kameo::message::{Context, Message};
-use l2_core::network::connection::HandleOutboundPacket;
 use l2_core::shared_packets::common::ReadablePacket;
 use l2_core::shared_packets::read::ReadablePacketBuffer;
-use std::time::Duration;
-use tokio::time::interval;
 use tracing::{info, instrument, warn};
 
 /// Helper function to calculate distance between two 3D points
@@ -96,81 +91,17 @@ impl Message<RequestMoveToLocation> for PlayerClient {
         }
         
         // Start or restart movement
-        let (mut movement, source_x, source_y, source_z) = self.start_movement(msg.x_to, msg.y_to, msg.z_to)?;
+        let (source_x, source_y, source_z) = self.start_movement(msg.x_to, msg.y_to, msg.z_to, ctx.actor_ref().clone())?;
         
-
-
-        // Spawn periodic broadcast task
-        let actor_ref = ctx.actor_ref().clone();
-        let controller = self.controller.clone();
-        let dest_x = msg.x_to;
-        let dest_y = msg.y_to;
-        let dest_z = msg.z_to;
-
-        let task_handle = tokio::spawn(async move {
-            let mut interval = tokio::time::interval(std::time::Duration::from_millis(300));
-            
-            loop {
-                // Check movement state BEFORE waiting
-                let result = actor_ref
-                    .ask(GetMovementPosition)
-                    .await;
-
-                if let Ok(Some((x, y, z, has_arrived))) = result {
-                    if has_arrived {
-                        // Arrived at destination, stop broadcasting
-                        info!("Player arrived at destination ({}, {}, {})", dest_x, dest_y, dest_z);
-                        break;
-                    }
-
-                    // Broadcast current position
-                    if let Ok(player) = actor_ref.ask(crate::pl_client::GetCharInfo).await {
-                        if let Ok(packet) = CharMoveToLocation::new(&player, dest_x, dest_y, dest_z) {
-                            controller.broadcast_packet(packet);
-                        }
-                    }
-                } else {
-                    // Error or no movement state, stop task
-                    break;
-                }
-
-                // Wait for next tick
-                interval.tick().await;
-            }
-        });
-
-        // Store the task handle in movement state and set it in client
-        movement.task_handle = Some(task_handle);
-        self.movement_state = Some(movement);
+        // Send initial movement packet immediately (optional, but good for responsiveness if task has delay)
+        // Note: The task sends update immediately, but we might want to send the "start" packet here?
+        // Actually, the previous logic removed the explicit broadcast.
+        // But wait, the task sends `CharMoveToLocation`.
+        // If I remove the explicit broadcast here, the task will send it.
+        // So I just need to call start_movement.
 
         Ok(())
     }
 }
 
-/// Message to get current movement position and status
-#[derive(Debug)]
-pub struct GetMovementPosition;
 
-impl Message<GetMovementPosition> for PlayerClient {
-    type Reply = anyhow::Result<Option<(i32, i32, i32, bool)>>;
-
-    async fn handle(
-        &mut self,
-        _: GetMovementPosition,
-        _ctx: &mut Context<Self, Self::Reply>,
-    ) -> Self::Reply {
-        if let Some(movement) = self.movement_state.as_ref() {
-            let (x, y, z) = movement.calculate_current_position();
-            let has_arrived = movement.has_arrived();
-            
-            // Update player's actual position
-            if let Ok(player) = self.try_get_selected_char_mut() {
-                let _ = player.set_location(x, y, z);
-            }
-            
-            Ok(Some((x, y, z, has_arrived)))
-        } else {
-            Ok(None)
-        }
-    }
-}
